Arreglada la impresión y la busqueda por cod articulo

+++ 4prendas/CapaNegocio/Negocio.cs
@@ -120,7 +120,7 @@ namespace CapaNegocio
             return datos.getTodosProductos();
         }
 
-        public List<Producto> getProdsCodigoBarras(int codigoBarras)
+        public List<Producto> getProdsCodigoBarras(long codigoBarras)
         {
             return datos.getProdsCodigoBarras(codigoBarras);
         }
